+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/feature-managers/transfer-point-feature-manager.ts
@@ -86,14 +86,14 @@ export class TransferPointFeatureManager extends MoveableFeatureManager<Transfer
     );
 
     public override onFeatureDrop(
-        droppedElement: Element,
+        droppedElement: Element | undefined,
         droppedOnFeature: Feature<Point>,
         dropEvent?: TranslateEvent
     ) {
         // TODO: droppedElement isn't necessarily a transfer point -> fix getElementFromFeature typings
         const droppedOnTransferPoint = this.getElementFromFeature(
             droppedOnFeature
-        ) as TransferPoint;
+        ) as TransferPoint | undefined;
         if (!droppedElement || !droppedOnTransferPoint) {
             console.error('Could not find element for the features');
             return false;