+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
     -   Publishing, accepting and marking radiograms as done.
     -   Accepting or denying resource request radiograms.
     -   Vehicle transfer, loading, unloading and occupation changes.
+    -   Visible status changes of patients.
+    -   Treating personnel changes for patients.
 -   By clicking on a log entry or a chart, a marker will be shown in the chart at that time. The log entry list scrolls to that time.
 
 ### Fixed